// src/types/graphql/generated.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  ABIJson: any;
  AppId: any;
  BlockchainData: any;
  BroadcastId: any;
  ChainId: any;
  ChallengeId: any;
  ContentEncryptionKey: any;
  CreateHandle: any;
  Cursor: any;
  DateTime: any;
  EncryptableDateTime: any;
  EncryptableMarkdown: any;
  EncryptableString: any;
  EncryptableTxHash: any;
  EncryptableURI: any;
  EncryptedPath: any;
  Ens: any;
  EvmAddress: any;
  Handle: any;
  ImageSizeTransform: any;
  Jwt: any;
  Locale: any;
  Markdown: any;
  MimeType: any;
  MomokaId: any;
  MomokaProof: any;
  NftGalleryId: any;
  NftGalleryName: any;
  Nonce: any;
  OnchainPublicationId: any;
  PoapEventId: any;
  ProfileId: any;
  PublicationId: any;
  Signature: any;
  TokenId: any;
  TxHash: any;
  TxId: any;
  URI: any;
  URL: any;
  UUID: any;
  UnixTimestamp: any;
  Void: any;
};
// ...
export type App = {
  __typename?: "App";
  id: Scalars["AppId"];
};
// ...
export type ArticleMetadataV3 = {
  __typename?: "ArticleMetadataV3";
  appId?: Maybe<Scalars["AppId"]>;
  attachments?: Maybe<Array<PublicationMetadataMedia>>;
  attributes?: Maybe<Array<MetadataAttribute>>;
  content: Scalars["EncryptableMarkdown"];
  contentWarning?: Maybe<PublicationContentWarningType>;
  encryptedWith?: Maybe<PublicationMetadataEncryptionStrategy>;
  hideFromFeed: Scalars["Boolean"];
  id: Scalars["String"];
  locale: Scalars["Locale"];
  marketplace?: Maybe<MarketplaceMetadata>;
  rawURI: Scalars["URI"];
  tags?: Maybe<Array<Scalars["String"]>>;
  /** The title of the article. Empty if not set. */
  title: Scalars["String"];
};

export type Asset = Erc20;

export type Audio = {
  __typename?: "Audio";
  mimeType?: Maybe<Scalars["MimeType"]>;
  uri: Scalars["URI"];
};

export type AudioMetadataV3 = {
  __typename?: "AudioMetadataV3";
  appId?: Maybe<Scalars["AppId"]>;
  asset: PublicationMetadataMediaAudio;
  attachments?: Maybe<Array<PublicationMetadataMedia>>;
  attributes?: Maybe<Array<MetadataAttribute>>;
  /** Optional content. Empty if not set. */
  content: Scalars["EncryptableMarkdown"];
  contentWarning?: Maybe<PublicationContentWarningType>;
  encryptedWith?: Maybe<PublicationMetadataEncryptionStrategy>;
  hideFromFeed: Scalars["Boolean"];
  id: Scalars["String"];
  locale: Scalars["Locale"];
  marketplace?: Maybe<MarketplaceMetadata>;
  rawURI: Scalars["URI"];
  tags?: Maybe<Array<Scalars["String"]>>;
  /** The title of the audio. Empty if not set. */
  title: Scalars["String"];
};
// ...
export type FollowModule =
  | FeeFollowModuleSettings
  | RevertFollowModuleSettings
  | UnknownFollowModuleSettings;
// ...
export type HandleInfo = {
  __typename?: "HandleInfo";
  /** The full handle - namespace/localname */
  fullHandle: Scalars["Handle"];
  guardian: HandleGuardianResult;
  /** The handle nft token id */
  id: Scalars["TokenId"];
  /** If null its not linked to anything */
  linkedTo?: Maybe<HandleLinkedTo>;
  /** The localname */
  localName: Scalars["String"];
  /** The namespace */
  namespace: Scalars["String"];
  ownedBy: Scalars["EvmAddress"];
  /** The suggested format to use on UI for ease but you can innovate and slice and dice as you want */
  suggestedFormatted: SuggestedFormattedHandle;
};
// ...
export type MarketplaceMetadata = {
  __typename?: "MarketplaceMetadata";
  animationUrl?: Maybe<Scalars["URI"]>;
  attributes?: Maybe<Array<PublicationMarketplaceMetadataAttribute>>;
  description?: Maybe<Scalars["Markdown"]>;
  externalURL?: Maybe<Scalars["URL"]>;
  image?: Maybe<ImageSet>;
  name?: Maybe<Scalars["String"]>;
};

export enum MarketplaceMetadataAttributeDisplayType {
  Date = "DATE",
  Number = "NUMBER",
  String = "STRING",
}

export type MentionNotification = {
  __typename?: "MentionNotification";
  id: Scalars["UUID"];
  publication: PrimaryPublication;
};

export type MetadataAttribute = {
  __typename?: "MetadataAttribute";
  key: Scalars["String"];
  /**
   * The type of the attribute. When:
   * - BOOLEAN: the `value` is `true`|`false`
   * - DATE: the `value` is a valid ISO 8601 date string
   * - NUMBER: the `value` is a valid JS number as string
   * - STRING: the `value` is a string.
   * - JSON: the `value` is a valid JSON serialized as string
   *
   */
  type: MetadataAttributeType;
  /** The value serialized as string. It's consumer responsibility to parse it according to `type`. */
  value: Scalars["String"];
};
// ...
export type MomokaInfo = {
  __typename?: "MomokaInfo";
  proof: Scalars["MomokaProof"];
};
// ...
export type NetworkAddress = {
  __typename?: "NetworkAddress";
  address: Scalars["EvmAddress"];
  chainId: Scalars["ChainId"];
};
// ...
export type OpenActionModule =
  | LegacyAaveFeeCollectModuleSettings
  | LegacyErc4626FeeCollectModuleSettings
  | LegacyFeeCollectModuleSettings
  | LegacyFreeCollectModuleSettings
  | LegacyLimitedFeeCollectModuleSettings
  | LegacyLimitedTimedFeeCollectModuleSettings
  | LegacyMultirecipientFeeCollectModuleSettings
  | LegacyRevertCollectModuleSettings
  | LegacySimpleCollectModuleSettings
  | LegacyTimedFeeCollectModuleSettings
  | MultirecipientFeeCollectOpenActionSettings
  | SimpleCollectOpenActionSettings
  | UnknownOpenActionModuleSettings;
// ...
export type Post = {
  __typename?: "Post";
  by: Profile;
  createdAt: Scalars["DateTime"];
  hashtagsMentioned: Array<Scalars["String"]>;
  id: Scalars["PublicationId"];
  isEncrypted: Scalars["Boolean"];
  isHidden: Scalars["Boolean"];
  metadata: PublicationMetadata;
  momoka?: Maybe<MomokaInfo>;
  openActionModules: Array<OpenActionModule>;
  operations: PublicationOperations;
  profilesMentioned: Array<ProfileMentioned>;
  publishedOn?: Maybe<App>;
  referenceModule?: Maybe<ReferenceModule>;
  stats: PublicationStats;
  txHash?: Maybe<Scalars["TxHash"]>;
};

export type PostStatsArgs = {
  request?: InputMaybe<PublicationStatsInput>;
};

export type PrfResult = {
  __typename?: "PrfResult";
  dd: Scalars["Boolean"];
  ss: Scalars["Boolean"];
};

export type PrimaryPublication = Comment | Post | Quote;

/** The Profile */
export type Profile = {
  __typename?: "Profile";
  /** When the profile was created */
  createdAt: Scalars["DateTime"];
  /** The follow module */
  followModule?: Maybe<FollowModule>;
  /** The profile follow nft address */
  followNftAddress?: Maybe<NetworkAddress>;
  guardian?: Maybe<ProfileGuardianResult>;
  /** The profile handle - a profile may not have one */
  handle?: Maybe<HandleInfo>;
  /** The profile id */
  id: Scalars["ProfileId"];
  interests: Array<Scalars["String"]>;
  invitedBy?: Maybe<Profile>;
  /** The number of invites left */
  invitesLeft: Scalars["Int"];
  /** The profile metadata. You can optionally query profile metadata by app id.  */
  metadata?: Maybe<ProfileMetadata>;
  /** The on chain identity */
  onchainIdentity: ProfileOnchainIdentity;
  operations: ProfileOperations;
  /** Who owns the profile */
  ownedBy: NetworkAddress;
  /** If the profile has been recommended by the authenticated user */
  peerToPeerRecommendedByMe: Scalars["Boolean"];
  /** If the profile has got signless enabled */
  signless: Scalars["Boolean"];
  /** If lens API will sponsor this persons for gasless experience, note they can have signless on but sponsor false which means it be rejected */
  sponsor: Scalars["Boolean"];
  stats: ProfileStats;
  txHash: Scalars["TxHash"];
};
// ...
export type ProfileGuardianResult = {
  __typename?: "ProfileGuardianResult";
  cooldownEndsOn?: Maybe<Scalars["DateTime"]>;
  protected: Scalars["Boolean"];
};
// ...
export type ProfileMentioned = {
  __typename?: "ProfileMentioned";
  profile: Profile;
  snapshotHandleMentioned: HandleInfo;
  stillOwnsHandle: Scalars["Boolean"];
};

export type ProfileMetadata = {
  __typename?: "ProfileMetadata";
  /** The app that this metadata is displayed on */
  appId?: Maybe<Scalars["AppId"]>;
  /** Profile Custom attributes */
  attributes?: Maybe<Array<MetadataAttribute>>;
  /** The bio for the profile */
  bio?: Maybe<Scalars["Markdown"]>;
  /** The cover picture for the profile */
  coverPicture?: Maybe<ImageSet>;
  /** The display name for the profile */
  displayName?: Maybe<Scalars["String"]>;
  /** The picture for the profile */
  picture?: Maybe<ProfilePicture>;
  /** The raw uri for the which the profile metadata was set as */
  rawURI: Scalars["URI"];
};

export type ProfileMirrorResult = {
  __typename?: "ProfileMirrorResult";
  mirrorId: Scalars["PublicationId"];
  mirroredAt: Scalars["DateTime"];
  profile: Profile;
};

export type ProfileOnchainIdentity = {
  __typename?: "ProfileOnchainIdentity";
  /** The ens information */
  ens?: Maybe<EnsOnchainIdentity>;
  /** The POH status */
  proofOfHumanity: Scalars["Boolean"];
  /** The sybil dot org information */
  sybilDotOrg: SybilDotOrgIdentity;
  /** The worldcoin identity */
  worldcoin: WorldcoinIdentity;
};

export type ProfileOperations = {
  __typename?: "ProfileOperations";
  canBlock: Scalars["Boolean"];
  canFollow: TriStateValue;
  canUnblock: Scalars["Boolean"];
  canUnfollow: Scalars["Boolean"];
  hasBlockedMe: OptimisticStatusResult;
  id: Scalars["ProfileId"];
  isBlockedByMe: OptimisticStatusResult;
  isFollowedByMe: OptimisticStatusResult;
  isFollowingMe: OptimisticStatusResult;
};
// ...
export type ProfileStats = {
  __typename?: "ProfileStats";
  comments: Scalars["Int"];
  countOpenActions: Scalars["Int"];
  followers: Scalars["Int"];
  following: Scalars["Int"];
  id: Scalars["ProfileId"];
  /** The profile classifier score of this profile relative to others on Lens. It is a % out of 100. */
  lensClassifierScore?: Maybe<Scalars["Float"]>;
  mirrors: Scalars["Int"];
  posts: Scalars["Int"];
  publications: Scalars["Int"];
  quotes: Scalars["Int"];
  /** How many times a profile has reacted on something */
  reacted: Scalars["Int"];
  /** How many times other profiles have reacted on something this profile did */
  reactions: Scalars["Int"];
};
// ...
export enum PublicationContentWarningType {
  Nsfw = "NSFW",
  Sensitive = "SENSITIVE",
  Spoiler = "SPOILER",
}

export type PublicationMarketplaceMetadataAttribute = {
  __typename?: "PublicationMarketplaceMetadataAttribute";
  displayType?: Maybe<MarketplaceMetadataAttributeDisplayType>;
  traitType?: Maybe<Scalars["String"]>;
  value?: Maybe<Scalars["String"]>;
};

export type PublicationMetadata =
  | ArticleMetadataV3
  | AudioMetadataV3
  | CheckingInMetadataV3
  | EmbedMetadataV3
  | EventMetadataV3
  | ImageMetadataV3
  | LinkMetadataV3
  | LiveStreamMetadataV3
  | MintMetadataV3
  | SpaceMetadataV3
  | StoryMetadataV3
  | TextOnlyMetadataV3
  | ThreeDMetadataV3
  | TransactionMetadataV3
  | VideoMetadataV3;

export type PublicationMetadataContentWarningFilter = {
  oneOf: Array<PublicationContentWarningType>;
};

export type PublicationMetadataEncryptionStrategy =
  PublicationMetadataLitEncryption;
// ...
export type PublicationMetadataMedia =
  | PublicationMetadataMediaAudio
  | PublicationMetadataMediaImage
  | PublicationMetadataMediaVideo;
// ...
export type PublicationMetadataMediaVideo = {
  __typename?: "PublicationMetadataMediaVideo";
  /** Alternative text for the video */
  altTag?: Maybe<Scalars["EncryptableString"]>;
  attributes?: Maybe<Array<MetadataAttribute>>;
  cover?: Maybe<EncryptableImageSet>;
  duration?: Maybe<Scalars["Int"]>;
  license?: Maybe<PublicationMetadataLicenseType>;
  video: EncryptableVideoSet;
};

export type PublicationMetadataTagsFilter = {
  all?: InputMaybe<Array<Scalars["String"]>>;
  oneOf?: InputMaybe<Array<Scalars["String"]>>;
};

export enum PublicationMetadataTransactionType {
  Erc20 = "ERC20",
  Erc721 = "ERC721",
  Other = "OTHER",
}

export type PublicationNotInterestedRequest = {
  on: Scalars["PublicationId"];
};

export type PublicationOperations = {
  __typename?: "PublicationOperations";
  actedOn: Array<OpenActionResult>;
  canAct: TriStateValue;
  canComment: TriStateValue;
  canDecrypt: CanDecryptResponse;
  canMirror: TriStateValue;
  canQuote: TriStateValue;
  hasActed: OptimisticStatusResult;
  hasBookmarked: Scalars["Boolean"];
  hasMirrored: Scalars["Boolean"];
  hasQuoted: Scalars["Boolean"];
  hasReacted: Scalars["Boolean"];
  hasReported: Scalars["Boolean"];
  id: Scalars["PublicationId"];
  isNotInterested: Scalars["Boolean"];
};
// ...
export type PublicationStats = {
  __typename?: "PublicationStats";
  bookmarks: Scalars["Int"];
  comments: Scalars["Int"];
  countOpenActions: Scalars["Int"];
  id: Scalars["PublicationId"];
  mirrors: Scalars["Int"];
  quotes: Scalars["Int"];
  reactions: Scalars["Int"];
};
// ...
export type ReferenceModule =
  | DegreesOfSeparationReferenceModuleSettings
  | FollowOnlyReferenceModuleSettings
  | LegacyDegreesOfSeparationReferenceModuleSettings
  | LegacyFollowOnlyReferenceModuleSettings
  | UnknownReferenceModuleSettings;
// ...
export type VideoMetadataV3 = {
  __typename?: "VideoMetadataV3";
  appId?: Maybe<Scalars["AppId"]>;
  asset: PublicationMetadataMediaVideo;
  attachments?: Maybe<Array<PublicationMetadataMedia>>;
  attributes?: Maybe<Array<MetadataAttribute>>;
  /** Optional content. Empty if not set. */
  content: Scalars["EncryptableMarkdown"];
  contentWarning?: Maybe<PublicationContentWarningType>;
  encryptedWith?: Maybe<PublicationMetadataEncryptionStrategy>;
  hideFromFeed: Scalars["Boolean"];
  id: Scalars["String"];
  isShortVideo: Scalars["Boolean"];
  locale: Scalars["Locale"];
  marketplace?: Maybe<MarketplaceMetadata>;
  rawURI: Scalars["URI"];
  tags?: Maybe<Array<Scalars["String"]>>;
  /** The title of the video. Empty if not set. */
  title: Scalars["String"];
};